Split structure.py into fragment.py and topology.py
- Created fragment.py with Fragment class
- Created topology.py with Topology class
- Removed structure.py
- Updated all imports across the codebase
